Move `SettingsListener` to service package
