+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -6,10 +6,12 @@ data class ServiceInstance(
     val daemon: MullvadDaemon,
     val connectionProxy: ConnectionProxy,
     val connectivityListener: ConnectivityListener,
-    val locationInfoCache: LocationInfoCache
+    val locationInfoCache: LocationInfoCache,
+    val settingsListener: SettingsListener
 ) {
     fun onDestroy() {
         connectionProxy.onDestroy()
         locationInfoCache.onDestroy()
+        settingsListener.onDestroy()
     }
 }
