+++ android/src/main/kotlin/net/mullvad/mullvadvpn/ui/ServiceConnection.kt
@@ -4,7 +4,6 @@ import net.mullvad.mullvadvpn.dataproxy.AccountCache
 import net.mullvad.mullvadvpn.dataproxy.AppVersionInfoCache
 import net.mullvad.mullvadvpn.dataproxy.KeyStatusListener
 import net.mullvad.mullvadvpn.dataproxy.RelayListListener
-import net.mullvad.mullvadvpn.dataproxy.SettingsListener
 import net.mullvad.mullvadvpn.service.ServiceInstance
 
 class ServiceConnection(private val service: ServiceInstance, val mainActivity: MainActivity) {
@@ -12,9 +11,9 @@ class ServiceConnection(private val service: ServiceInstance, val mainActivity: 
     val connectionProxy = service.connectionProxy
     val connectivityListener = service.connectivityListener
     val locationInfoCache = service.locationInfoCache
+    val settingsListener = service.settingsListener
 
     val keyStatusListener = KeyStatusListener(daemon)
-    val settingsListener = SettingsListener(daemon)
     val appVersionInfoCache = AppVersionInfoCache(mainActivity, daemon, settingsListener)
     val accountCache = AccountCache(settingsListener, daemon)
     var relayListListener = RelayListListener(daemon, settingsListener)
